requirements/dev.txt
-r base.txt

drf-yasg==1.20.0
django-debug-toolbar==3.2.4
